requirements.txt
Django==5.0.10
djangorestframework==3.14.0
django-cors-headers==3.13.0
pyodbc==5.0.0
